+++ tempodb/backend/azure/azure_helpers_test.go
@@ -5,6 +5,8 @@ import (
 	"os"
 	"testing"
 
+	"github.com/Azure/go-autorest/autorest/adal"
+	"github.com/Azure/go-autorest/autorest/azure"
 	"github.com/grafana/dskit/flagext"
 	"github.com/stretchr/testify/assert"
 )
@@ -12,6 +14,8 @@ import (
 const (
 	TestStorageAccountName = "foobar"
 	TestStorageAccountKey  = "abc123"
+	TestAzureClientId      = "myClientId"
+	TestAzureTenantId      = "myTenantId"
 )
 
 // TestGetStorageAccountName* explicitly broken out into
@@ -108,4 +112,42 @@ func TestGetContainerURL(t *testing.T) {
 			assert.Equal(t, tc.expectedURL, url.String())
 		})
 	}
+}
+
+func TestServicePrincipalTokenFromFederatedToken(t *testing.T) {
+	os.Setenv("AZURE_CLIENT_ID", TestAzureClientId)
+	defer os.Unsetenv("AZURE_CLIENT_ID")
+	os.Setenv("AZURE_TENANT_ID", TestAzureTenantId)
+	defer os.Unsetenv("AZURE_TENANT_ID")
+
+	mockOAuthConfig, _ := adal.NewOAuthConfig("foo", "bar")
+	mockedServicePrincipalToken := new(adal.ServicePrincipalToken)
+
+	tmpDir := t.TempDir()
+	_ = os.WriteFile(tmpDir+"/jwtToken", []byte("myJwtToken"), 0666)
+	os.Setenv("AZURE_FEDERATED_TOKEN_FILE", tmpDir+"/jwtToken")
+	defer os.Unsetenv("AZURE_FEDERATED_TOKEN_FILE")
+
+	newOAuthConfigFunc := func(activeDirectoryEndpoint, tenantID string) (*adal.OAuthConfig, error) {
+		assert.Equal(t, azure.PublicCloud.ActiveDirectoryEndpoint, activeDirectoryEndpoint)
+		assert.Equal(t, TestAzureTenantId, tenantID)
+
+		_, err := adal.NewOAuthConfig(activeDirectoryEndpoint, tenantID)
+		assert.NoError(t, err)
+
+		return mockOAuthConfig, nil
+	}
+
+	servicePrincipalTokenFromFederatedTokenFunc := func(oauthConfig adal.OAuthConfig, clientID string, jwt string, resource string, callbacks ...adal.TokenRefreshCallback) (*adal.ServicePrincipalToken, error) {
+		assert.True(t, *mockOAuthConfig == oauthConfig, "should return the mocked object")
+		assert.Equal(t, TestAzureClientId, clientID)
+		assert.Equal(t, "myJwtToken", jwt)
+		assert.Equal(t, "https://bar.blob.core.windows.net", resource)
+		return mockedServicePrincipalToken, nil
+	}
+
+	token, err := servicePrincipalTokenFromFederatedToken("AzureGlobal", "https://bar.blob.core.windows.net", newOAuthConfigFunc, servicePrincipalTokenFromFederatedTokenFunc)
+
+	assert.NoError(t, err)
+	assert.True(t, mockedServicePrincipalToken == token, "should return the mocked object")
 }